+++ cmd/wharf/vars.go
@@ -5,7 +5,9 @@ import (
 
 	"github.com/iver-wharf/wharf-cmd/internal/flagtypes"
 	"github.com/iver-wharf/wharf-cmd/internal/lastbuild"
+	"github.com/iver-wharf/wharf-cmd/pkg/wharfyml"
 	"github.com/spf13/cobra"
+	"gopkg.in/typ.v4/slices"
 )
 
 var varsFlags = struct {
@@ -38,4 +40,17 @@ func init() {
 	addBuildIDFlag(varsCmd.PersistentFlags(), &varsFlags.buildID)
 	addWharfYmlEnvFlag(varsCmd, varsCmd.PersistentFlags(), &varsFlags.env)
 	addWharfYmlInputsFlag(varsCmd, varsCmd.PersistentFlags(), &varsFlags.inputs)
+}
+
+func varsCmdParseBuildDef(args []string) (wharfyml.Definition, error) {
+	currentDir, err := parseCurrentDir(slices.SafeGet(args, 0))
+	if err != nil {
+		return wharfyml.Definition{}, err
+	}
+
+	return parseBuildDefinition(currentDir, wharfyml.Args{
+		Env:       varsFlags.env,
+		Inputs:    parseInputArgs(varsFlags.inputs),
+		VarSource: newBuildIDVarSource(varsFlags.buildID),
+	})
 }
