+++ cmd/wharf/vars_list.go
@@ -41,16 +41,7 @@ in multiple sources.`,
 		return []string{"yml"}, cobra.ShellCompDirectiveFilterFileExt
 	},
 	RunE: func(cmd *cobra.Command, args []string) error {
-		currentDir, err := parseCurrentDir(slices.SafeGet(args, 0))
-		if err != nil {
-			return err
-		}
-
-		def, err := parseBuildDefinition(currentDir, wharfyml.Args{
-			Env:       varsFlags.env,
-			Inputs:    parseInputArgs(varsFlags.inputs),
-			VarSource: newBuildIDVarSource(varsFlags.buildID),
-		})
+		def, err := varsCmdParseBuildDef(args)
 		if err != nil {
 			return err
 		}
